+++ tests/opentelemetry-test-utils/src/opentelemetry/test/metrictestutil.py
@@ -18,6 +18,7 @@ from typing import Optional
 from opentelemetry.attributes import BoundedAttributes
 from opentelemetry.sdk.metrics.export import (
     AggregationTemporality,
+    DataPointFlags,
     Gauge,
     Histogram,
     HistogramDataPoint,
@@ -50,9 +51,12 @@ def _generate_sum(
     description=None,
     unit=None,
     is_monotonic=True,
+    flags=None,
 ) -> Metric:
     if attributes is None:
         attributes = BoundedAttributes(attributes={"a": 1, "b": True})
+    if flags is None:
+        flags = DataPointFlags.get_default()
     return _generate_metric(
         name,
         Sum(
@@ -62,6 +66,7 @@ def _generate_sum(
                     start_time_unix_nano=1641946015139533244,
                     time_unix_nano=1641946016139533244,
                     value=value,
+                    flags=flags,
                 )
             ],
             aggregation_temporality=AggregationTemporality.CUMULATIVE,
@@ -73,10 +78,17 @@ def _generate_sum(
 
 
 def _generate_gauge(
-    name, value, attributes=None, description=None, unit=None
+    name,
+    value,
+    attributes=None,
+    description=None,
+    unit=None,
+    flags=None,
 ) -> Metric:
     if attributes is None:
         attributes = BoundedAttributes(attributes={"a": 1, "b": True})
+    if flags is None:
+        flags = DataPointFlags.get_default()
     return _generate_metric(
         name,
         Gauge(
@@ -86,6 +98,7 @@ def _generate_gauge(
                     start_time_unix_nano=None,
                     time_unix_nano=1641946016139533244,
                     value=value,
+                    flags=flags,
                 )
             ],
         ),
@@ -110,9 +123,12 @@ def _generate_histogram(
     attributes: Attributes = None,
     description: Optional[str] = None,
     unit: Optional[str] = None,
+    flags: Optional[DataPointFlags] = None,
 ) -> Metric:
     if attributes is None:
         attributes = BoundedAttributes(attributes={"a": 1, "b": True})
+    if flags is None:
+        flags = DataPointFlags.get_default()
     return _generate_metric(
         name,
         Histogram(
@@ -127,6 +143,7 @@ def _generate_histogram(
                     explicit_bounds=[123.0, 456.0],
                     min=1,
                     max=457,
+                    flags=flags,
                 )
             ],
             aggregation_temporality=AggregationTemporality.CUMULATIVE,